move directory order
