+++ app/src/main/java/com/jin/honey/feature/paymentdetail/ui/PaymentDetailViewModel.kt
@@ -2,7 +2,7 @@ package com.jin.honey.feature.paymentdetail.ui
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.jin.domain.model.order.Order
+import com.jin.domain.order.model.Order
 import com.jin.domain.usecase.GetOrderDetailUseCase
 import com.jin.feature.ui.state.UiState
 import kotlinx.coroutines.flow.MutableStateFlow
